Update List-chromeExtensions.ps1
**1. Extensions are missing from the "Default" directory.**
After doing a few bits of troubleshooting, I figured out this issue. The key pivot point of the script was to get the names of the extension from their manifest.json file and put it into the hash table. Well it seems that not all extensions have this information in it in a human readable way (or blatantly misleading). In my testing I found a number of apps falling through the cracks due to this; however, they seem to be Google directly related? Either way, I took the block of code out that checked for these "bad" names.
As a result, we are now missing a way to know which names are for what extension. So I added a new item in the hash table that shows where the manifest.json file is loaded from.

2. Multiple profiles on the same host enhancement
This one was rather interesting in how to properly enumerate all of the profiles without causing undue stress on the host. The simplest way to handle this was to loosen the initial "Default query" and to add in a wild card. Now this does do a lookup on the host for over a thousand objects but it seems to only have a minimal impact (no CPU/RAM increase; just a slight delay in results).